Addressed all code review issues.  Almost a complete rewrite of all     functionality.
Changed SlugInput to FilteringTextInput, a general version of the same
    functionality that subclasses TextInput directly instead of FormInput.

Changed SlugifyingInput to PrePopulatingTextInput, a general version of the same
    functionality.

Removed VerifiedPasswordInput and will create a new branch for it.

Removed an uneccessary class (TogglingCheckboxInput) that I hacked together
    before I realized the dependency system existed.

Improved all Python and JavaScript docstrings.

TODO: add unit tests for all new classes.